Fix integer division
This was making all of the ranks 0, because the numbers being divided
were ints. After they were cast to doubles, this appeared to fix that
issue.

+++ src/main/java/edu/byu/cstaheli/cs453/suggesting_similar_queries/rank/SuggestedQueryRanker.java
@@ -84,7 +84,7 @@ public class SuggestedQueryRanker
      */
     private double getOccurrenceOfSuggestedQuery()
     {
-        return trie.frequency(suggestedQuery) / trie.getMostCommonQueryFrequency();
+        return (double) trie.frequency(suggestedQuery) / (double) trie.getMostCommonQueryFrequency();
     }
 
     /**
@@ -124,7 +124,7 @@ public class SuggestedQueryRanker
      */
     private double getNumberOfDirectModifications()
     {
-        return trie.getFrequencyOfAdjacency(originalQuery, suggestedQuery) / trie.getMostCommonQueryFrequency();
+        return (double) trie.getFrequencyOfAdjacency(originalQuery, suggestedQuery) / (double) trie.getMostCommonQueryFrequency();
     }
 
     public double getRank()
